refactor: 로또 결과를 담을 LottoResult 클래스 생성
